+++ app/src/main/java/net/androidbootcamp/myapplication/MainActivity.java
@@ -1,32 +1,43 @@
 package net.androidbootcamp.myapplication;
 
+import android.content.Intent;
 import android.os.Bundle;
+import android.view.View;
 import android.widget.Button;
-import android.widget.TextView;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 public class MainActivity extends AppCompatActivity {
-TextView username_input;
-TextView password_input;
-Button signin_button;
-
+    Button signIn_button;
+    Button register_button;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
-//--------------------------------------------------
-//Sign in button
-//---------------------------------------------------
-       signin_button = findViewById(R.id.submit_button);
+        //------------------
+        // Sign In
+        //------------------
+        signIn_button = findViewById(R.id.signin_button2);
+        signIn_button.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Intent inToMain = new Intent(MainActivity.this, Login.class);
+                startActivity(inToMain);
+            }
+        });
 
-//----------------------------------------------------
-// User Sign In Input
-//----------------------------------------------------
-username_input = findViewById(R.id.username_input);
-
-password_input = findViewById(R.id.password_input);
+        //-----------------------
+        // Register
+        //-----------------------
+        register_button = findViewById(R.id.register_button);
+        register_button.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Intent inToMain = new Intent(MainActivity.this, Register_Vendor.class);
+                startActivity(inToMain);
+            }
+        });
     }
 }
