+++ app/src/main/java/net/androidbootcamp/myapplication/Vendor_Items.java
@@ -4,11 +4,11 @@ import android.os.Bundle;
 
 import androidx.appcompat.app.AppCompatActivity;
 
-public class navigation_vendor extends AppCompatActivity {
+public class Vendor_Items extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_navigation_vendor);
+        setContentView(R.layout.activity_vendor_items);
     }
 }
