+++ app/src/main/res/layout/activity_main.xml
@@ -5,51 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#f7ffdd"
-    tools:context=".MainActivity" >
-
-    <EditText
-        android:id="@+id/password_input"
-        android:layout_width="250dp"
-        android:layout_height="50dp"
-        android:layout_marginBottom="68dp"
-        android:background="#FFFFFF"
-        android:ems="10"
-        android:hint="PASSWORD"
-        android:inputType="textPassword"
-        android:textSize="16sp"
-        app:layout_constraintBottom_toTopOf="@+id/submit_button"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.496"
-        app:layout_constraintStart_toStartOf="parent" />
-
-    <Button
-        android:id="@+id/submit_button"
-        android:layout_width="145dp"
-        android:layout_height="40dp"
-        android:background="#B1CF5F"
-        android:text="SIGN IN"
-        android:textColor="#707070"
-        android:textSize="16sp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.707" />
-
-    <EditText
-        android:id="@+id/username_input"
-        android:layout_width="250dp"
-        android:layout_height="50dp"
-        android:layout_marginTop="10dp"
-        android:background="#FFFFFF"
-        android:ems="10"
-        android:hint="USERNAME"
-        android:inputType="textPassword"
-        android:textSize="16sp"
-        app:layout_constraintBottom_toTopOf="@+id/password_input"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/phone_icon" />
+    tools:context=".MainActivity">
 
     <ImageView
         android:id="@+id/phone_icon"
@@ -68,18 +24,32 @@
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintVertical_bias="0.068" />
 
-    <TextView
-        android:id="@+id/textView2"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="8dp"
-        android:text="Forgot Password?"
-        android:textStyle="bold"
-        app:layout_constraintBottom_toTopOf="@+id/submit_button"
+    <Button
+        android:id="@+id/signin_button2"
+        android:layout_width="145dp"
+        android:layout_height="40dp"
+        android:layout_marginTop="100dp"
+        android:background="#B1CF5F"
+        android:text="SIGN IN"
+        android:textColor="#707070"
+        android:textSize="16sp"
+        app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.727"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/password_input"
-        app:layout_constraintVertical_bias="0.0" />
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintVertical_bias="0.483" />
+
+    <Button
+        android:id="@+id/register_button"
+        android:layout_width="145dp"
+        android:layout_height="40dp"
+        android:layout_marginBottom="124dp"
+        android:background="#B1cf5f"
+        android:text="REGISTER"
+        android:textColor="#707070"
+        android:textSize="16sp"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
