+++ app/src/main/res/layout/activity_navigation_vendor.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".navigation_vendor">
+    tools:context=".Navigation_Vendor">
 
     <TextView
         android:id="@+id/textView4"
