+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,4 @@
 <resources>
     <string name="app_name">My Application</string>
+    <string name="forgot_password">Forgot Password?</string>
 </resources>
